requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
psycopg2-binary
python-dotenv==1.0.0
google-generativeai==0.8.6
groq>=0.9.0
openai>=1.0.0
requests
beautifulsoup4
pdfplumber>=0.10.0
python-docx>=1.1.0
gunicorn